Migrate conftest to e2e_tests folder
